Simulating a web service to provide an alternative definition.
If the user attempts to search for a word which doesn't exist, a web
service from another server could help us to get an alternative
definition. Here, we're just simulating that behaviour.

It's interesting to find out that the web service would return a
future if we want to work asynchronously. However, if we want to do
so, we have to make some modifications in our service. Specifically,
we have to use the 'Action.async' which works directly on Futures.