Naughty word request is forbidden.
However, the solution is not made of 'ActionBuilder's because I'm
finding it very difficult to compose an 'Action.async' with the
existing 'WordFilter'. Perhaps, this will be the next task to be
done.

On the other hand, the filter is not fully correct. For example, if I
write 'fuckencio' as part of the definition of a new word, it will be
ignored. This is not the right behaviour, it should deny the whole
words instead.