The word transducer knows how to deal with the WS response.
Before, we were using an Option[String] as a result of the web service
invocation. This is not right, since a web service should return a
`Future` instead. Therefore, I've changed the `ws` method signature in
order to keep the rest of elements aware of it. The word transducer
had to migrate to the new signature and it has received a considerable
refactoring.