Can't use copy method in WrappedRequest.
I've moved `WordRequest` to a case class, aiming to use the `copy`
method for a nicer syntax. However, the copy method wasn't compiling,
because it wasn't able to detect the `definition` field. I've been
investigating about the issue and I've find out that a case class
which contains or inherits a copy method DOES NOT override it. While
wandering by the `WrappedRequest` api I could see that the method does
already exist, and therefore the copy logic is not the one I was
expecting.

Since the case class let me remove the object apply method, I've
decided to keep it anyway.