Removed 'side effect' from search chain.
I noticed that we were introducing innecessary side effects in the
code. I've created a StateTransformer which turns a normal request
into a state request (a request which contains the state). By doing
so, any function which needs to update or consult the state is able to
do so in a local way.