Initial users have been moved to the FakeApplication.
And it seems to work! I've been wandering by the API and I found that
the Cache is just a plugin. By overriding the plugin list of the
application and by setting an EhCache with some values, all tests are
working fine. There's a lot of work to do yet, but undoubtedly, it's a
big step.