I've provided some ways to provide users and words while testing.
Now, if you want to declare some users and words, you could make use
of the tools `withUsers` and `withWords`. Both rely on the new `reset`
methods which I've added in the singletons which deal with the cache.

Besides, I've added `withoutUsers` and `withoutWords`. They are
useless in the standing situation, but we could set default values for
all tests in the future. Given that scenario, these tools could be
used to remove the default users and words.